requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.1
Werkzeug==3.0.3
PyJWT==2.8.0
python-dotenv==1.0.1